requirements.txt
PyPDF2==3.0.1
openai==1.40.3
faiss-cpu==1.8.0.post1
numpy==1.26.4
slack_sdk==3.31.0
flask==3.0.3
python-dotenv==1.0.1